moved new file stats stuff into own package
