+++ data-model-solr/src/main/java/org/mousephenotype/cda/file/stats/StatsJson.java
@@ -1,4 +1,4 @@
-package org.mousephenotype.cda.solr.service;
+package org.mousephenotype.cda.file.stats;
 
 public class StatsJson {
